requirements.txt
PyGithub
GitPython
python-pushover